+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## DEV
 
  - Turtle personality makes more initial fabbers at each tech level
+ - Integrated wondible's Section of Foreign Intelligence for Galactic War
 
 ## 0.13.1 - 2019-08-30